# docs/api/openapi.json
{"openapi":"3.0.1","info":{"title":"Digital Therapy Assistant API","description":"REST API for burnout support, CBT sessions, diary entries, progress tracking, and crisis workflows.","license":{"name":"Academic Use"},"version":"0.1.0"},"servers":[{"url":"http://127.0.0.1:8080","description":"Generated server url"}],"paths":{"/crisis/safety-plan":{"get":{"tags":["crisis-controller"],"operationId":"getSafetyPlan","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SafetyPlanResponse"}}}}}},"put":{"tags":["crisis-controller"],"operationId":"updateSafetyPlan","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateSafetyPlanRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SafetyPlanResponse"}}}}}}},"/sessions/{sessionId}/start":{"post":{"tags":["session-controller"],"operationId":"startSession","parameters":[{"name":"sessionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StartSessionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}}}}},"/sessions/{sessionId}/end":{"post":{"tags":["session-controller"],"operationId":"endSession","parameters":[{"name":"sessionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}},{"name":"reason","in":"query","required":false,"schema":{"type":"string","default":"completed"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}}}}},"/sessions/{sessionId}/chat":{"post":{"tags":["session-controller"],"operationId":"chat","parameters":[{"name":"sessionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChatRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ChatResponse"}}}}}}},"/diary/entries":{"get":{"tags":["diary-controller"],"operationId":"getEntries","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DiaryEntryResponse"}}}}}}},"post":{"tags":["diary-controller"],"operationId":"createEntry","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDiaryEntryRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiaryEntryResponse"}}}}}}},"/diary/distortions/suggest":{"post":{"tags":["diary-controller"],"operationId":"suggest","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DistortionSuggestionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ThoughtAnalysisResponse"}}}}}}},"/crisis/detect":{"post":{"tags":["crisis-controller"],"operationId":"detect","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DistortionSuggestionRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CrisisResponse"}}}}}}},"/auth/register":{"post":{"tags":["auth-controller"],"operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/refresh":{"post":{"tags":["auth-controller"],"operationId":"refresh","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/auth/logout":{"post":{"tags":["auth-controller"],"operationId":"logout","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefreshRequest"}}},"required":true},"responses":{"200":{"description":"OK"}}}},"/auth/login":{"post":{"tags":["auth-controller"],"operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AuthResponse"}}}}}}},"/sessions":{"get":{"tags":["session-controller"],"operationId":"listSessions","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SessionResponse"}}}}}}}},"/sessions/{sessionId}":{"get":{"tags":["session-controller"],"operationId":"getSession","parameters":[{"name":"sessionId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/SessionResponse"}}}}}}},"/progress/weekly":{"get":{"tags":["progress-controller"],"operationId":"weekly","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProgressResponse"}}}}}}},"/progress/monthly":{"get":{"tags":["progress-controller"],"operationId":"monthly","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProgressResponse"}}}}}}},"/progress/burnout":{"get":{"tags":["progress-controller"],"operationId":"burnout","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ProgressResponse"}}}}}}},"/progress/achievements":{"get":{"tags":["progress-controller"],"operationId":"achievements","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/AchievementsResponse"}}}}}}},"/diary/insights":{"get":{"tags":["diary-controller"],"operationId":"getInsights","parameters":[{"name":"userId","in":"query","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiaryInsightsResponse"}}}}}}},"/diary/entries/{entryId}":{"get":{"tags":["diary-controller"],"operationId":"getEntry","parameters":[{"name":"entryId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DiaryEntryResponse"}}}}}},"delete":{"tags":["diary-controller"],"operationId":"deleteEntry","parameters":[{"name":"entryId","in":"path","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"OK"}}}},"/crisis":{"get":{"tags":["crisis-controller"],"operationId":"crisisState","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CrisisResponse"}}}}}}},"/crisis/coping-strategies":{"get":{"tags":["crisis-controller"],"operationId":"getCopingStrategies","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/CopingStrategiesResponse"}}}}}}}},"components":{"schemas":{"UpdateSafetyPlanRequest":{"required":["safetyPlan","userId"],"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"safetyPlan":{"maxLength":5000,"minLength":0,"type":"string"}}},"SafetyPlanResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"safetyPlan":{"type":"string"}}},"StartSessionRequest":{"required":["userId"],"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"title":{"maxLength":255,"minLength":0,"type":"string"},"mood":{"maxLength":100,"minLength":0,"type":"string"},"mode":{"maxLength":100,"minLength":0,"type":"string"}}},"SessionResponse":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"title":{"type":"string"},"status":{"type":"string"},"startedAt":{"type":"string","format":"date-time"},"endedAt":{"type":"string","format":"date-time"},"summary":{"type":"string"},"endReason":{"type":"string"}}},"ChatRequest":{"required":["message"],"type":"object","properties":{"message":{"maxLength":4000,"minLength":0,"type":"string"}}},"ChatResponse":{"type":"object","properties":{"sessionId":{"type":"string","format":"uuid"},"userMessage":{"type":"string"},"assistantMessage":{"type":"string"},"crisisFlagged":{"type":"boolean"},"timestamp":{"type":"string","format":"date-time"}}},"CreateDiaryEntryRequest":{"required":["automaticThought","emotion","reflection","userId"],"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"automaticThought":{"maxLength":3000,"minLength":0,"type":"string"},"emotion":{"maxLength":255,"minLength":0,"type":"string"},"reflection":{"maxLength":4000,"minLength":0,"type":"string"}}},"DiaryEntryResponse":{"type":"object","properties":{"id":{"type":"string","format":"uuid"},"userId":{"type":"string","format":"uuid"},"automaticThought":{"type":"string"},"emotion":{"type":"string"},"reflection":{"type":"string"},"suggestedDistortions":{"type":"array","items":{"type":"string"}},"createdAt":{"type":"string","format":"date-time"}}},"DistortionSuggestionRequest":{"required":["text"],"type":"object","properties":{"text":{"maxLength":3000,"minLength":0,"type":"string"}}},"ThoughtAnalysisResponse":{"type":"object","properties":{"distortions":{"type":"array","items":{"type":"string"}},"reframingPrompts":{"type":"array","items":{"type":"string"}}}},"CrisisResponse":{"type":"object","properties":{"crisis":{"type":"boolean"},"riskLevel":{"type":"string"},"keywordsDetected":{"type":"array","items":{"type":"string"}},"action":{"type":"string"},"reasoning":{"type":"string"}}},"RegisterRequest":{"required":["email","fullName","password"],"type":"object","properties":{"fullName":{"maxLength":100,"minLength":2,"type":"string"},"email":{"type":"string"},"password":{"maxLength":100,"minLength":8,"type":"string"}}},"AuthResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"fullName":{"type":"string"},"email":{"type":"string"},"accessToken":{"type":"string"},"refreshToken":{"type":"string"},"tokenType":{"type":"string"}}},"RefreshRequest":{"required":["refreshToken"],"type":"object","properties":{"refreshToken":{"type":"string"}}},"LoginRequest":{"required":["email","password"],"type":"object","properties":{"email":{"type":"string"},"password":{"type":"string"}}},"ProgressResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"timeframe":{"type":"string"},"score":{"type":"number","format":"double"},"completedSessions":{"type":"integer","format":"int32"},"diaryEntries":{"type":"integer","format":"int32"},"crisisAlerts":{"type":"integer","format":"int32"},"burnoutTrend":{"type":"string"},"highlights":{"type":"array","items":{"type":"string"}}}},"AchievementsResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"completedSessions":{"type":"integer","format":"int32"},"diaryEntries":{"type":"integer","format":"int32"},"achievements":{"type":"array","items":{"type":"string"}},"nextMilestone":{"type":"string"}}},"DiaryInsightsResponse":{"type":"object","properties":{"userId":{"type":"string","format":"uuid"},"entryCount":{"type":"integer","format":"int32"},"topDistortions":{"type":"array","items":{"type":"string"}},"reframingPrompts":{"type":"array","items":{"type":"string"}},"insightSummary":{"type":"string"}}},"CopingStrategiesResponse":{"type":"object","properties":{"strategies":{"type":"array","items":{"type":"string"}},"emergencyResources":{"type":"array","items":{"type":"string"}},"note":{"type":"string"}}}}}}